Refactored base Test fixtures
Refactored base implementation of test fixtures and made it more simpler
to write the code in test case itself. Also in earlier case for each
`in` clause will execute the scan every time for sub sections of the
unit test. With the updated implementation, now the scan will be invoked
only once for each `should` section.